# archive/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: VibeCode WebGUI API
  version: 1.0.0
  description: |
    Comprehensive API for VibeCode WebGUI - a development platform with AI-powered coding assistance,
    workspace management, and collaborative development tools.

    ## Security Features
    - CSRF Protection with double-submit cookie pattern
    - Rate limiting with Redis/Valkey backend
    - JWT-based authentication with NextAuth.js
    - Multi-factor authentication (TOTP, SMS, Email)
    - Request validation with Zod schemas
    - Security headers and CORS protection

    ## Authentication
    The API supports multiple authentication methods:
    - OAuth2 (GitHub, Google)
    - Credentials-based with secure password hashing
    - JWT tokens for session management
    - CSRF tokens for state-changing operations

    ## Rate Limiting
    Different endpoints have different rate limits:
    - Strict: 5 requests per minute
    - Auth: 10 requests per 5 minutes
    - API: 100 requests per minute (bypassed for authenticated users)
    - Upload: 5 requests per 5 minutes

  contact:
    name: VibeCode Team
    url: https://vibecode.dev
    email: support@vibecode.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://vibecode.dev/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
  - CookieAuth: []
  - CSRFToken: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from NextAuth.js authentication
    
    CookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session cookie from NextAuth.js
    
    CSRFToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: CSRF token for state-changing operations (required with cookie auth)
  schemas:
...
    Workspace:
      type: object
      required:
        - id
        - projectName
        - framework
        - status
      properties:
        id:
          type: string
          pattern: ^[a-zA-Z0-9_-]+$
          description: Unique workspace identifier
        projectId:
          type: string
          description: Associated project identifier
        projectName:
          type: string
          maxLength: 200
          description: Human-readable project name
        framework:
          type: string
          maxLength: 50
          description: Development framework (e.g., 'react', 'nextjs', 'python')
        userId:
          type: string
          description: Owner user ID
        status:
          type: string
          enum: [creating, active, stopped, error, deleting]
          description: Current workspace status
        files:
          type: object
          additionalProperties:
            type: string
          description: File system structure (path -> content mapping)
        dependencies:
          type: array
          items:
            type: string
          description: Package dependencies
        environment:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /ai/chat/stream:
    post:
      tags: [AI, Chat]
      summary: Stream AI chat response
      description: |
        Sends a message to AI and receives a streaming response using Server-Sent Events.
        Optimized for real-time conversations with immediate feedback.
      security:
        - BearerAuth: []
        - CookieAuth: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - conversationId
                - message
              properties:
                conversationId:
                  type: string
                  maxLength: 100
                  description: Conversation identifier
                message:
                  type: string
                  minLength: 1
                  maxLength: 100000
                  description: User message
                model:
                  type: string
                  default: "anthropic/claude-3.5-sonnet"
                  description: AI model to use
                workspaceId:
                  type: string
                  default: "default"
                  description: Workspace context
                files:
                  type: array
                  items:
                    type: string
                  default: []
                  description: Files to include in context
                enableWebSearch:
                  type: boolean
                  default: false
                  description: Enable web search capability
                enableRAG:
                  type: boolean
                  default: true
                  description: Enable retrieval-augmented generation
      responses:
        '200':
          description: Streaming response initiated
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-sent events with AI response chunks

  # Workspace Management
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces
      description: Retrieves a paginated list of user workspaces
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [creating, active, stopped, error, deleting]
          description: Filter by workspace status
      responses:
        '200':
          description: Workspaces retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                  total:
                    type: integer
                    description: Total number of workspaces
                  page:
                    type: integer
                    description: Current page number
                  limit:
                    type: integer
                    description: Items per page
                  hasNext:
                    type: boolean
                    description: Whether there are more pages

    post:
      tags: [Workspaces]
      summary: Create workspace
      description: |
        Creates a new development workspace with the specified configuration.
        Provisions resources in Kubernetes cluster if available.
      security:
        - BearerAuth: []
        - CookieAuth: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - projectName
                - framework
                - files
              properties:
                projectId:
                  type: string
                  maxLength: 100
                  description: Unique project identifier
                projectName:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Human-readable project name
                framework:
                  type: string
                  minLength: 1
                  maxLength: 50
                  description: Development framework
                userId:
                  type: string
                  default: "anonymous"
                  description: Owner user ID
                files:
                  type: object
                  additionalProperties:
                    type: string
                  description: Initial file structure
                dependencies:
                  type: array
                  items:
                    type: string
                  default: []
                  description: Package dependencies
                environment:
                  type: object
                  additionalProperties:
                    type: string
                  default: {}
                  description: Environment variables
      responses:
        '201':
          description: Workspace created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Invalid request parameters
        '401':
          description: Authentication required
        '503':
          description: Kubernetes cluster not available
...